requirements.txt
Django==3.1.3
djangorestframework==3.12.2
django-cors-headers==3.5.0
psycopg2==2.8
pytest==6.2.5
pytest-django==3.8.0
six==1.16.0
python-memcached==1.59.0